Functioning Sudoku Solver without question 5 text
Functions make_move, save_board, is_complete and solve_board with helper function recursive_solve have been added to solve questions 1-4. For question 5, analysis has to be written out, but a counter for the number of times the recursive function has been called is included. Still need to decide whether to count the make_move function or the recursive function.